fixed tick

+++ assignment_package/src/mygl.h
@@ -29,7 +29,7 @@ private:
     InputBundle m_inputs; // A collection of variables to be updated in keyPressEvent, mouseMoveEvent, mousePressEvent, etc.
 
     QTimer m_timer; // Timer linked to tick(). Fires approximately 60 times per second.
-    float m_currTime; // Current time used to compute dT for player movement
+    double m_currTime; // Current time used to compute dT for player movement
 
     void moveMouseToCenter(); // Forces the mouse position to the screen's center. You should call this
                               // from within a mouse move event after reading the mouse movement so that
